+++ src/dazzle/cli/runtime_impl/__init__.py
@@ -16,6 +16,7 @@ The modules are:
 from .build import (
     build_api_command,
     build_command,
+    build_css_command,
     build_ui_command,
     migrate_command,
 )
@@ -29,6 +30,7 @@ __all__ = [
     # Build commands
     "build_ui_command",
     "build_api_command",
+    "build_css_command",
     "migrate_command",
     "build_command",
     # Serve